refactor(tests): Replace `RequestProvider` with `ServerParamsPsr7Provider` in data provider annotations for improved test organization and clarity.
